requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
google-genai
python-dotenv
python-multipart
